requirements.txt
# Core Dependencies
torch>=1.10.0
numpy>=1.21.0
matplotlib>=3.4.0
pandas>=1.3.0

# For Kaggle download
kaggle>=1.5.0
python-dotenv>=0.19.0
